+++ R/constraints.R
@@ -37,6 +37,10 @@ NULL
 #'     ensure that all selected planning units have at least a certain number
 #'     of neighbors.}
 #'
+#'   \item{[add_boundary_constraints()]}{
+#'     Add constraints to a
+#'     ensure that selected planning units are not overly spatially fragmented.}
+#'
 #'   \item{[add_contiguity_constraints()]}{
 #'     Add constraints to a
 #'     ensure that all selected planning units are spatially connected to each
@@ -92,17 +96,20 @@ NULL
 #' # create problem with neighbor constraints
 #' p4 <- p1 %>% add_neighbor_constraints(2)
 #'
+#' # create problem with boundary constraints
+#' p5 <- p1 %>% add_boundary_constraints(0.5)
+#'
 #' # create problem with contiguity constraints
-#' p5 <- p1 %>% add_contiguity_constraints()
+#' p6 <- p1 %>% add_contiguity_constraints()
 #'
 #' # create problem with feature contiguity constraints
-#' p6 <- p1 %>% add_feature_contiguity_constraints()
+#' p7 <- p1 %>% add_feature_contiguity_constraints()
 #'
 #' # solve problems
-#' s <- terra::rast(lapply(list(p1, p2, p3, p4, p5, p6), solve))
+#' s <- terra::rast(lapply(list(p1, p2, p3, p4, p5, p6, p7), solve))
 #' names(s) <- c(
 #'   "minimal problem", "locked in", "locked out",
-#'   "neighbor", "contiguity", "feature contiguity"
+#'   "neighbor", "boundary", "contiguity", "feature contiguity"
 #' )
 #'
 #' # plot solutions