+++ R/write_problem.R
@@ -14,7 +14,10 @@ NULL
 #' to specify whether the problem formulation will be saved in the
 #' [LP](https://docs.gurobi.com/current/#refman/lp_format.html) or
 #' [MPS](https://docs.gurobi.com/current/#refman/mps_format.html)
-#' format (respectively).
+#' format (respectively). If using the Gurobi solver
+#' (i.e., `solver = "gurobi"`), the problem can
+#' be saved in a compressed file format (e.g., `".mps.gx"`) to reduce
+#' file size (see `?gurobi::gurobi_write()` for details).
 #'
 #' @param solver `character` name of optimization solver to write the problem
 #' to disk. Available options include: `"rsymphony" `, `"gurobi"`, or `NULL`.
@@ -67,14 +70,6 @@ write_problem <- function(x, path, solver = NULL) {
     assertthat::noNA(path),
     assertthat:: is.writeable(dirname(path))
   )
-  assert(
-    assertthat::has_extension(path, "lp") ||
-    assertthat::has_extension(path, "mps"),
-    msg = paste(
-      "{.arg path} must have a",
-      "{.val \".lp\"} or {.val \".mps\"} file extension."
-    )
-  )
   if (is.null(solver)) {
     solver <- "rsymphony"
     if (requireNamespace("gurobi", quietly = TRUE)) {
@@ -96,7 +91,35 @@ write_problem <- function(x, path, solver = NULL) {
   if (identical(solver, "gurobi")) {
     assert(is_installed("gurobi")) # nocov
   }
-
+  # assert valid file extensions
+  valid_ext <- c(".lp", ".mps")
+  valid_ext_msg <- paste(
+    "{.arg path} must have a",
+    "{.val \".lp\"} or {.val \".mps\"} file extension for the Rsymphony",
+    "solver."
+  )
+  if (identical(solver, "gurobi")) {
+    valid_ext <- c(
+      ".mps", ".rew", ".lp", ".rlp", ".dua", ".dlp", ".ilp",
+      ".mps.gz", ".rew.gz", ".lp.gz", ".rlp.gz", ".dua.gz", ".dlp.gz", "ilp.gz",
+      ".mps.bz2", ".rew.bz2", ".lp.bz2", ".rlp.bz2", ".dua.bz2", ".dlp.bz2",
+      "ilp.bz2",
+      ".mps.zip", ".rew.zip", ".lp.zip", ".rlp.zip", ".dua.zip", ".dlp.zip",
+      "ilp.zip",
+      ".mps.7z", ".rew.7z", ".lp.7z", ".rlp.7z", ".dua.7z", ".dlp.7z", "ilp.7z"
+    )
+    valid_ext_msg <- c(
+      "!" = paste(
+        "{.arg path} must have a valid file extension",
+        "for the Gurobi solver."
+      ),
+      "i" = "See {.fn gurobi::gurobi_write} for valid file extensions."
+    )
+  }
+  assert(
+    isTRUE(any(vapply(valid_ext, endsWith, logical(1), x = path))),
+    msg = valid_ext_msg
+  )
   # add solver
   if (identical(solver, "rsymphony")) {
     x <- suppressWarnings(add_rsymphony_solver(x))