+++ src/init.c
@@ -12,6 +12,7 @@ extern SEXP _prioritizr_rcpp_absolute_amount_held_by_solution(SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_add_rij_data(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_add_zones_constraints(SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_apply_asym_connectivity_penalties(SEXP, SEXP, SEXP);
+extern SEXP _prioritizr_rcpp_apply_boundary_constraints(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_apply_boundary_penalties(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_apply_bounded_constraints(SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_apply_connectivity_penalties(SEXP, SEXP, SEXP);
@@ -31,6 +32,7 @@ extern SEXP _prioritizr_rcpp_apply_min_penalties_objective(SEXP, SEXP, SEXP, SEX
 extern SEXP _prioritizr_rcpp_apply_min_set_objective(SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_apply_min_shortfall_objective(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_apply_neighbor_constraints(SEXP, SEXP, SEXP);
+extern SEXP _prioritizr_rcpp_apply_surrounded_penalties(SEXP, SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_asym_connectivity(SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_boundary(SEXP, SEXP, SEXP, SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_branch_matrix(SEXP);
@@ -66,6 +68,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_prioritizr_rcpp_add_rij_data",                                      (DL_FUNC) &_prioritizr_rcpp_add_rij_data,                                      4},
     {"_prioritizr_rcpp_add_zones_constraints",                             (DL_FUNC) &_prioritizr_rcpp_add_zones_constraints,                             2},
     {"_prioritizr_rcpp_apply_asym_connectivity_penalties",                 (DL_FUNC) &_prioritizr_rcpp_apply_asym_connectivity_penalties,                 3},
+    {"_prioritizr_rcpp_apply_boundary_constraints",                        (DL_FUNC) &_prioritizr_rcpp_apply_boundary_constraints,                        7},
     {"_prioritizr_rcpp_apply_boundary_penalties",                          (DL_FUNC) &_prioritizr_rcpp_apply_boundary_penalties,                          7},
     {"_prioritizr_rcpp_apply_bounded_constraints",                         (DL_FUNC) &_prioritizr_rcpp_apply_bounded_constraints,                         5},
     {"_prioritizr_rcpp_apply_connectivity_penalties",                      (DL_FUNC) &_prioritizr_rcpp_apply_connectivity_penalties,                      3},
@@ -85,6 +88,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_prioritizr_rcpp_apply_min_set_objective",                           (DL_FUNC) &_prioritizr_rcpp_apply_min_set_objective,                           3},
     {"_prioritizr_rcpp_apply_min_shortfall_objective",                     (DL_FUNC) &_prioritizr_rcpp_apply_min_shortfall_objective,                     4},
     {"_prioritizr_rcpp_apply_neighbor_constraints",                        (DL_FUNC) &_prioritizr_rcpp_apply_neighbor_constraints,                        3},
+    {"_prioritizr_rcpp_apply_surrounded_penalties",                        (DL_FUNC) &_prioritizr_rcpp_apply_surrounded_penalties,                        4},
     {"_prioritizr_rcpp_asym_connectivity",                                 (DL_FUNC) &_prioritizr_rcpp_asym_connectivity,                                 2},
     {"_prioritizr_rcpp_boundary",                                          (DL_FUNC) &_prioritizr_rcpp_boundary,                                          6},
     {"_prioritizr_rcpp_branch_matrix",                                     (DL_FUNC) &_prioritizr_rcpp_branch_matrix,                                     1},